requirements.txt
coverage
Django
django-filter
django-taggit
djangorestframework
model-mommy
Pillow
django-ckeditor
django-crispy-forms
mysqlclient
six
